refactor: move constants to utils
